Added patrolling NPC
Its path is assigned with several 'target' gameobjects, represented by
yellow markers set not to render in the game itself.

The player can push it around a little bit, the extent to which this
happens can be adjusted with the NPC's rigidbody mass.

If time allows it may be a good idea to have the NPC react to the
player; right now it just casually pushes the player along or out of the
way.