Created Pentagram and InGameGUI.cs
Not much code is actually testable currently, a way to trigger the
addItem() function of the pentagram when an item is set down is needed.
The InGameGUI contains a gold counter that should be increased based on
the 'value' (specified in Pentagram.cs) of an item that gets destroyed.
I have GameManager set to create the InGameGUI from a prefab when a
StartRound() void is called, but this can easily be changed.

Additionally, I coded a neededItemQuantities check, but if the round is
won based on gold instead this can be deleted.