Added item property script, speed modifier
Created rudimentary item prefabs (and shitty sprites) for each of the 3
test items. Added an array for storing raw material quantities in the
InGameGUI script. A specified speed modifier is now applied to the
player while they hold an item. Lastly, fixed (?) an apparent bug with
the possibleItems array containing multiple of the same object.