Added a lot of NPC code
Added a bit more complexity to NPC behavior, so they turn and look at
the player whenever they see the player while suspicious, see an item
held by the player, or are touched by the player when suspicious. The
obstacle avoidance also got a couple improvements.

I added a system to change vision cone colors, we'll have to see if this
fits the game design. The system also required that the player's
collider and item triggers be split into separate gameobjects, in order
to be on different layers. I created two new prefabs, witch2 and
patroller2.